+++ Interop/ONNX/src/test/java/org/tribuo/interop/onnx/extractors/BERTFeatureExtractorTest.java
@@ -1,4 +1,4 @@
-package org.tribuo.interop.onnx.bert;
+package org.tribuo.interop.onnx.extractors;
 
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
@@ -10,7 +10,6 @@ import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.List;
-import java.util.Map;
 
 public class BERTFeatureExtractorTest {
